# Core static request refactor, getting rid of response handler