--- a/packages/libs/core/src/handle/index.ts
+++ b/packages/libs/core/src/handle/index.ts
@@ -2,3 +2,4 @@ export * from "./api";
 export * from "./default";
 export * from "./fallback";
 export { getCustomHeaders, setCustomHeaders } from "./headers";
+export * from "./types";