--- a/packages/libs/lambda-at-edge/src/default-handler.ts
+++ b/packages/libs/lambda-at-edge/src/default-handler.ts
@@ -137,6 +137,7 @@ const getS3File = (
       bucketName,
       buildId,
       file,
+      isData,
       region
     });