This implements my initial version of the changes discussed in #1144 – moving more of the logic to core and getting rid of the response handler.

Most default handler tests remain unported (I disabled them to run e2e tests) and I'll look into moving them to under libs/core where possible. The unlocalized, no basepath tests I updated while developing this.

A couple of questions:
1. Is the complexity of some pages using S3 client and some origin requests worth it?
2. What to do with public files and custom headers? The alternatives I can think of are to say "unsupported", to use S3 client and break any large files with headers or to add a response handler conditionally.