@jvarho 2021-06-13
There are a lot of e2e test failures, which wasn't unexpected. Some are clear enough – e.g. some methods have to be disallowed in the request handler. Others less so. I'll have a look in the next few days.