@dphang 2021-06-18
> This implements my initial version of the changes discussed in #1144 – moving more of the logic to core and getting rid of the response handler.
> 
> Most default handler tests remain unported (I disabled them to run e2e tests) and I'll look into moving them to under libs/core where possible. The unlocalized, no basepath tests I updated while developing this.
> 
> A couple of questions:
> 
> 1. Is the complexity of some pages using S3 client and some origin requests worth it?
> 2. What to do with public files and custom headers? The alternatives I can think of are to say "unsupported", to use S3 client and break any large files with headers or to add a response handler conditionally.

1. Yes I think pages should always use S3 client? Since they should almost always be < 1 MB.
2. Public files is in the grey area, I suppose most people may not be using custom headers on public files, but a few might have large files, so probably optimize for that? If they are using custom headers, probably we should say they are only supported if < 1 MB (we could add new cache behaviors specifically for public files with an origin response handler, but it might add too much complexity)