@jvarho 2021-06-19
@dphang There are still many test failures, but this works well enough that I benchmarked it.

I've compared this branch to current master. I'm connecting to eu-central-1 with the S3 bucket in us-east-1, to try to include that latency in the picture. I'm requesting a HTML page, the same page each time, but with a query parameter to prevent cloudfront caching.

**Before:**

1st request (cold start)

0.51s END - START
68 + 7 ms Billed Duration
44.6 + 1.14 ms handler execution time

2nd request (1 minute later)

0.54s END - START
79 + 23 ms Billed Duration
0.39 + 0.29 ms handler execution time

3rd request (immediate after 2nd)

0.40s END - START
2 + 2 ms Billed Duration
0.29 + 0.19 ms handler execution time


**After:**


1st request (cold start)

2.26s END - START
2259 ms Billed Duration
2201.7 ms handler execution time

2nd request (1 minute later)

0.82s END - START
819 Billed Duration
688.5 ms handler execution time

3rd request (immediate after 2nd)

0.12s END - START
117 Billed Duration
106.7 ms handler execution time


I'm not sure how reproducible these results are, but it looks like the cold start penalty is much worse and the billed duration always significantly larger. In the best case it can be faster, but I suspect there is some caching going on that may not happen normally – normally the same file would not be requested soon (since it should get cached in cloudfront) unless it is e.g. a 404 page.

Maybe I've done something stupid in the implementation or am missing something, but this does not look very encouraging to me.