@dphang 2021-06-19
Strange, let me try to test your changes too. I remember any AWS service called in Lambda is supposed to go over AWS's private networks as well, so using s3 client should be as fast as using an s3 origin.

When I tested it was a while back and with us-east-1 only, maybe that is affecting it?