@dphang 2021-06-22
@jvarho maybe one thing you can check is whether it is using the right regional S3. From S3 docs https://docs.aws.amazon.com/general/latest/gr/s3.html. Basically it should be `s3.us-east-1.amazonaws.com` (or other bucket). I think that it should be using the right regional endpoints, but maybe it's not?

Recently with some AWS permanent redirect failure causing s3 uploads at deploy-time, gave me this idea. Since that somehow was not using the right endpoint and I had to set it manually, although that was in AWS SDK JS v2