@dphang 2021-06-22
Nevermind, tested and it does seem to be using the right regional endpoints in AWS SDK JS v3. Though according to here it should go over AWS networks? https://aws.amazon.com/blogs/networking-and-content-delivery/leveraging-external-data-in-lambdaedge/

```
Leverage the AWS network by using AWS-based endpoints like Amazon S3, DynamoDB, API Gateway, ELB, and so on. When you connect to AWS resources, your application traffic travels on the high-quality network that Amazon monitors for both availability and low latency. This monitoring has the additional benefit of keeping error rates low and window sizes high.
```