@jvarho 2021-09-04
I'm not pursuing this, so closing. The branch will be there in case someone wants to work on it.